htmlOutputTemplate bug
After moving from 4.x to 5.1 I noticed that the html template stopped working. I pulled down the source and tracked down the issue. It looks like htmlOutputTemplate was missing from the delayTemplateExpansion array and grunt would replace the placeholders as it would try and resolve them (so your generated html.ts files would not have the html content). I looked to add a test but it looks like one was added but not being run correctly (it needed the html settings within the options object and the settings and expected template needed some tweaks to get the diff passing. As I was preparing this pull request I noticed someone else had also picked up on this issue as well.